passing in SVN_DIRENT_ALL was causing a TypeError exception in cmd_list. This is because SVN_DIRENT_ALL is 0xFFFFFFFF which does not fit in an int so use long. Add long support to arg processing.
git-svn-id: https://svn.code.sf.net/p/pysvn/code/trunk/pysvn/Extension@935 8a6430f3-a0d1-4a6a-aeff-73e05446c159

+++ Source/pysvn_client_cmd_list.cpp
@@ -268,7 +268,7 @@ Py::Object pysvn_client::cmd_list( const Py::Tuple &a_args, const Py::Dict &a_kw
 #else
     bool recurse = args.getBoolean( name_recurse, false );
 #endif
-    apr_uint32_t dirent_fields = args.getInteger( name_dirent_fields, SVN_DIRENT_ALL );
+    apr_uint32_t dirent_fields = args.getLong( name_dirent_fields, SVN_DIRENT_ALL );
     bool fetch_locks = args.getBoolean( name_fetch_locks, false );
 
     revisionKindCompatibleCheck( is_url, peg_revision, name_peg_revision, name_url_or_path );
